[sdk-make-qmltypes] Fix use with explicit shadow build
Some packages are known to use %build section like this:

%build
mkdir build && cd build
%qmake5 -r ..
make

This turned to be an issue when mb2 was updated so that it does not need
the %qtc_* macros for qmake/make isolation. Before that, mb2 would
execute qmake directly, not via rpmbuild, so the `cd build` step would
not be executed. After that, mb2 executes the %build section correctly,
and thus producing the Makefile under path different than the source
path.